pyproject.toml
[project]
name = "camel-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "camel-ai>=0.2.76",
]
